🎉 Add BasicList component for improved data display and interaction
This commit introduces the BasicList component, which enhances data presentation with sorting, mobile and desktop layouts, and optimized record handling. It replaces and refactors existing list code for better organization and user experience.
