+++ packages/ui/src/types/index.ts
@@ -1,2 +1,2 @@
 export type { MapPickerProps, DynamicFormProps, DanceEngineEntity, MetaData, DynamicFieldOptions } from "./form";   // ✅ Expose only selected types
-export type { BasicListProps } from "./list"
+export type { BasicListProps, BasicListColumnValueAdapter } from "./list"
